修改index生成方式 修改时间拼写错误。 升级到5.5

+++ GameCore/GameLogic/Game/AIBehaviorTree/ActionDistancTarget.cs
@@ -24,7 +24,7 @@ namespace GameLogic.Game.AIBehaviorTree
 				yield break;
 			}
 
-			var target = root.Perception.State[(long)index] as BattleCharacter;
+			var target = root.Perception.State[(int)index] as BattleCharacter;
 			if (target == null)
 			{
 				yield return RunStatus.Failure;
